Enable placeholder resolving for extra field declaration in gateway
* Moved common placeholders which are used in both edge services (gateway and connectivity)
  to a separate module
* Used those placeholders in gateway in SessionedSignal to resolve the
  extra field declaration
