+++ connectivity/api/src/main/java/org/eclipse/ditto/connectivity/api/EnforcementFactoryFactory.java
@@ -47,9 +47,9 @@ public final class EnforcementFactoryFactory {
      * Creates new instance of {@link EnforcementFilterFactory} that is preconfigured with the following placeholders
      * for the filters:
      * <ul>
-     * <li>{@link org.eclipse.ditto.connectivity.api.placeholders.ThingPlaceholder}</li>
-     * <li>{@link org.eclipse.ditto.connectivity.api.placeholders.PolicyPlaceholder}</li>
-     * <li>{@link org.eclipse.ditto.connectivity.api.placeholders.EntityIdPlaceholder}</li>
+     * <li>{@link org.eclipse.ditto.edge.api.placeholders.ThingPlaceholder}</li>
+     * <li>{@link org.eclipse.ditto.edge.api.placeholders.PolicyPlaceholder}</li>
+     * <li>{@link org.eclipse.ditto.edge.api.placeholders.EntityIdPlaceholder}</li>
      * </ul>
      *
      * @param <I> the type from which the input values are resolved
