+++ edge/api/src/main/java/org/eclipse/ditto/edge/api/placeholders/AbstractEntityIdPlaceholder.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.ditto.connectivity.api.placeholders;
+package org.eclipse.ditto.edge.api.placeholders;
 
 import java.util.Arrays;
 import java.util.Collections;
