+++ edge/api/src/main/java/org/eclipse/ditto/edge/api/placeholders/PolicyPlaceholder.java
@@ -10,17 +10,19 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.ditto.connectivity.api.placeholders;
+package org.eclipse.ditto.edge.api.placeholders;
 
-import org.eclipse.ditto.base.model.signals.Signal;
+import org.eclipse.ditto.base.model.entity.id.EntityId;
 import org.eclipse.ditto.placeholders.Placeholder;
 
 /**
  * A {@link org.eclipse.ditto.placeholders.Placeholder} that requires a {@code String}
- * (a valid Feature ID) to resolve its placeholders.
- * A {@link org.eclipse.ditto.placeholders.Placeholder} that requires a {@code String} (a valid Feature ID) to resolve its placeholders.
- *
- * @since 1.5.0
+ * (a valid Policy ID) to resolve its placeholders.
  */
-public interface FeaturePlaceholder extends Placeholder<Signal<?>> {
+public interface PolicyPlaceholder extends Placeholder<EntityId> {
+
+    static PolicyPlaceholder getInstance() {
+        return ImmutablePolicyPlaceholder.INSTANCE;
+    }
+
 }
