+++ edge/api/src/test/java/org/eclipse/ditto/edge/api/placeholders/ImmutableThingPlaceholderTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2017 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -10,11 +10,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.ditto.connectivity.api.placeholders;
+package org.eclipse.ditto.edge.api.placeholders;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
-import org.eclipse.ditto.policies.model.PolicyId;
 import org.eclipse.ditto.things.model.ThingId;
 import org.junit.Test;
 import org.mutabilitydetector.unittesting.MutabilityAssert;
@@ -24,22 +24,21 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 
 /**
- * Tests {@link ImmutableEntityIdPlaceholder}.
+ * Tests {@link ImmutableThingPlaceholder}.
  */
-public class ImmutableEntityPlaceholderTest {
+public class ImmutableThingPlaceholderTest {
 
     private static final String NAME = "ditto";
     private static final String NAMESPACE = "eclipse";
-    private static final PolicyId POLICY_ID = PolicyId.of(NAMESPACE, NAME);
     private static final ThingId THING_ID = ThingId.of(NAMESPACE, NAME);
-    private static final EntityIdPlaceholder UNDER_TEST = ImmutableEntityIdPlaceholder.INSTANCE;
+    private static final ThingPlaceholder UNDER_TEST = ImmutableThingPlaceholder.INSTANCE;
 
     /**
      * Assert immutability.
      */
     @Test
     public void assertImmutability() {
-        MutabilityAssert.assertInstancesOf(ImmutablePolicyPlaceholder.class, MutabilityMatchers.areImmutable());
+        MutabilityAssert.assertInstancesOf(ImmutableThingPlaceholder.class, MutabilityMatchers.areImmutable());
     }
 
     /**
@@ -47,15 +46,10 @@ public class ImmutableEntityPlaceholderTest {
      */
     @Test
     public void testHashCodeAndEquals() {
-        EqualsVerifier.forClass(ImmutablePolicyPlaceholder.class)
+        EqualsVerifier.forClass(ImmutableThingPlaceholder.class)
                 .suppress(Warning.INHERITED_DIRECTLY_FROM_OBJECT)
                 .usingGetClass()
                 .verify();
-    }
-
-    @Test
-    public void testReplacePolicyId() {
-        assertThat(UNDER_TEST.resolveValues(POLICY_ID, "id")).contains(POLICY_ID.toString());
     }
 
     @Test
@@ -64,18 +58,8 @@ public class ImmutableEntityPlaceholderTest {
     }
 
     @Test
-    public void testReplacePolicyName() {
-        assertThat(UNDER_TEST.resolveValues(POLICY_ID, "name")).contains(NAME);
-    }
-
-    @Test
     public void testReplaceThingName() {
         assertThat(UNDER_TEST.resolveValues(THING_ID, "name")).contains(NAME);
-    }
-
-    @Test
-    public void testReplacePolicyNamespace() {
-        assertThat(UNDER_TEST.resolveValues(POLICY_ID, "namespace")).contains(NAMESPACE);
     }
 
     @Test
@@ -85,7 +69,19 @@ public class ImmutableEntityPlaceholderTest {
 
     @Test
     public void testUnknownPlaceholderReturnsEmpty() {
-        assertThat(UNDER_TEST.resolveValues(POLICY_ID, "policy_id")).isEmpty();
+        assertThat(UNDER_TEST.resolveValues(THING_ID, "thing_id")).isEmpty();
+    }
+
+    @Test
+    public void testResolvingWithNull() {
+        assertThatExceptionOfType(NullPointerException.class)
+                .isThrownBy(() -> UNDER_TEST.resolveValues(THING_ID, null));
+    }
+
+    @Test
+    public void testResolvingWithEmptyString() {
+        assertThatExceptionOfType(IllegalArgumentException.class)
+                .isThrownBy(() -> UNDER_TEST.resolveValues(THING_ID, ""));
     }
 
 }
